requirements.txt
pinecone-client[grpc]
mmh3 
ipywidgets
gensim==4.0.1
sentence-transformers
datasets
tiktoken
openai
langchain
pythainlp
pinecone-datasets 
sentence-transformers
protobuf==3.20
googletrans==3.1.0a0